requirements.txt
openai==0.27.5
Pillow==9.5.0